fix(health-checks): import JobOwners without booting the dagster package
`JobOwners` is a plain enum, but it lived under `posthog/dags/`, whose `__init__`
calls `django.setup()`. Every non-dagster importer therefore re-applied LOGGING,
and `disable_existing_loggers` silenced every logger built up to that point.

Creating a team reaches that import through the eager health-check task, so any
test that did so left earlier modules unable to log, which broke `caplog`
assertions in unrelated suites.
