+++ products/data_warehouse/backend/temporal/health_checks/external_data_failure.py
@@ -3,7 +3,7 @@ import datetime as dt
 from django.db.models import Q
 from django.utils import timezone
 
-from posthog.dags.common.owners import JobOwners
+from posthog.job_owners import JobOwners
 from posthog.models.health_issue import HealthIssue
 from posthog.temporal.health_checks.detectors import DEFAULT_EXECUTION_POLICY
 from posthog.temporal.health_checks.framework import AlertContent, HealthCheck, Remediation
