+++ products/data_warehouse/backend/temporal/health_checks/materialized_view_failure.py
@@ -1,6 +1,6 @@
 from django.db.models import Q
 
-from posthog.dags.common.owners import JobOwners
+from posthog.job_owners import JobOwners
 from posthog.models.health_issue import HealthIssue
 from posthog.temporal.health_checks.detectors import DEFAULT_EXECUTION_POLICY
 from posthog.temporal.health_checks.framework import (
