+++ products/web_analytics/backend/temporal/health_checks/authorized_urls.py
@@ -1,6 +1,6 @@
 from django.db.models import Q
 
-from posthog.dags.common.owners import JobOwners
+from posthog.job_owners import JobOwners
 from posthog.models.health_issue import HealthIssue
 from posthog.models.team import Team
 from posthog.temporal.health_checks.detectors import DEFAULT_EXECUTION_POLICY
